refactor: bit 값을 long 타입으로 수정

+++ src/main/java/IdGenerator.java
@@ -23,8 +23,8 @@ public class IdGenerator {
                 timestampId,
                 shardId,
                 sequenceId,
-                (int) shardIdGenerator.allocatedBits(),
-                (int) sequenceIdGenerator.allocatedBits()
+                shardIdGenerator.allocatedBits(),
+                sequenceIdGenerator.allocatedBits()
         );
     }
